sparse matrix
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/diseasemapping@2316 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/gpuRandom/src/cholBatch.cpp
@@ -199,7 +199,7 @@ int cholBatchVcl(
   A.size2() * A.internal_size2(),// NpadBetweenMatrices,
   NlocalCache, 
   Nlocal,
-  A.size2() > NlocalCache[0]); // allow overflow
+  ((int) A.size2() ) > NlocalCache[0]); // allow overflow
 
   viennacl::ocl::context ctx(viennacl::ocl::get_context(ctx_id));
   viennacl::ocl::program & my_prog = ctx.add_program(cholClString, "my_kernel");
